deploy/element145-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.11.0
google-generativeai>=0.3.0
pydantic>=2.0.0
